parse: allow gnu getopt_long style behavior
GNU `getopt_long` allows for arguments and options to be intermixed, for
example "foo bar.c --help".  This questionable behavior can now be opted in
to.  We do this by sorting the given arguments when we see the first
non-option (in this example "bar.c") and shuffling options back to the
front of the argument list.